requirements.txt
# Основные зависимости для RAG-системы
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.29.0
faiss-cpu==1.7.4
tiktoken==0.5.2
openai==1.6.1

# Дополнительные зависимости для работы с данными
datasets==2.14.0
huggingface-hub==0.19.0
